requirements.txt
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
click==8.1.8
colorama==0.4.6
dynamodb==1.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pynamodb==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
